Treinando Debug

+++ _18 - Debugging/src/CalculadoraDeMedias.java
@@ -5,17 +5,17 @@ public class CalculadoraDeMedias {
         Scanner scan = new Scanner(System.in);
         String[] alunos = {"Lucas", "Lunah", "Ariana", "Hermes"};
 
-        int media = calculaMediaDaTurma(alunos, scan);
+        double media = calculaMediaDaTurma(alunos, scan);
 
-        System.out.printf("Media da turma %d" , media);
+        System.out.printf("Media da turma %.1f" , media);
     }
 
-    public static int calculaMediaDaTurma(String[] alunos, Scanner scanner) {
+    public static double calculaMediaDaTurma(String[] alunos, Scanner scanner) {
 
-        int soma = 0;
+        double soma = 0;
         for(String aluno : alunos) {
             System.out.printf("Nota do aluno %s: ", aluno);
-            int nota = scanner.nextInt();
+            double nota = scanner.nextInt();
             soma += nota;
         }
